move FastaSequenceIndexEntry to a separate file and make it public
